refactor(atomic): move result templates to shared folder
https://coveord.atlassian.net/browse/KIT-1529
