+++ packages/atomic/src/components/result-templates/atomic-result-template/atomic-result-template.tsx
@@ -1,7 +1,7 @@
 import {Component, Element, Prop, Method, State} from '@stencil/core';
 import {ResultTemplate, ResultTemplateCondition} from '@coveo/headless';
-import {MapProp} from '../../utils/props-utils';
-import {ResultTemplateCommon} from '../result-templates/result-template-common';
+import {MapProp} from '../../../utils/props-utils';
+import {ResultTemplateCommon} from '../result-template-common';
 
 /**
  * The `atomic-result-template` component determines the format of the query results, depending on the conditions that are defined for each template. A `template` element must be the child of an `atomic-result-template`, and either an `atomic-result-list` or `atomic-folded-result-list` must be the parent of each `atomic-result-template`.
